type converion added int to float data type

+++ type_conversion.c
@@ -4,8 +4,8 @@
 int main()
 {
     int a=99 , b=2;
-    float div_value = a/b;
-    printf("THE DIV VALUE IS : %f\n",div_value); // output = 49.000000
+    float div_value = (float) a/b; // type conversion integer to float data type
+    printf("THE DIV VALUE IS : %.1f\n",div_value); // output = 49.000000
     return 0;
 
 }
